requirements.txt
flask
pillow
requests
numpy
python-dotenv